App_falling/requirements.txt
# numpy==1.21.6
numpy==1.23.5
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
pyyaml==6.0
# PyQt5-sip==12.7.0
PyQt5-sip
# PyQt5==5.13.1
PyQt5
PyQtWebEngine
qt-thread-updater==1.1.6
Pillow==9.5.0
shapely==2.0.1
pyinstaller
psutil==5.9.8
imutils
tqdm
pandas
torch
torchvision
matplotlib
seaborn
scipy
requests